requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
djangorestframework==3.14.0
future==0.18.3
idna==3.4
iso8601==1.1.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
serial==0.0.97
sqlparse==0.4.4
urllib3==1.26.15
